[IMP] file-not-used: Ignore unused files into the import directory
If there are data files into the "import" directory, e.g. XML/CSV,
they are probably intended to be used by the import python script, hence
they won't be referenced from the manifest.

This covers the case when there is complicated structure, or validation
that needs to be imported from modules init hook, and file is delivered
along with the module.

This commit causes the lint to ignore files located into the
"import" directory.